+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Common:
 * Added dataset loader for Babel dataset.
 
+Task 1, Predict Modality:
+* Starter Kit: Added starter kit for Python users
+
 Task 3, Joint Embedding:
 * Metric: Added TotalVI metric 'latent_mixing_metric'.